# public/openapi.yaml
openapi: 3.0.0
info:
  title: Expense Tracker API
  description: A RESTful API for tracking personal expenses with user management and analytics
  version: 1.0.0
  contact:
    name: API Support
    email: support@expense-tracker.com
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.expense-tracker.com
    description: Production server
...
paths:
...
  /api/users:
    get:
      tags:
        - Users
      summary: Get all users
      description: Retrieve a list of all users, optionally filtered by email
      parameters:
        - name: email
          in: query
          description: Filter users by email address
          schema:
            type: string
            format: email
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  count:
                    type: integer
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Users
      summary: Create a new user
      description: Register a new user in the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
              properties:
                name:
                  type: string
                  example: John Doe
                email:
                  type: string
                  format: email
                  example: john@example.com
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Bad request (invalid data)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Users
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Bad request
        '404':
          description: User not found
        '409':
          description: Email already exists

    delete:
      tags:
        - Users
      summary: Delete user
      description: Delete a user and all associated expenses
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '404':
          description: User not found
...
  /api/webhooks:
    get:
      tags:
        - Webhooks
      summary: List all webhooks
      responses:
        '200':
          description: List of registered webhooks
          content:
            application/json:
              schema:
                type: object
                properties:
                  webhooks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Webhook'
                  count:
                    type: integer

    post:
      tags:
        - Webhooks
      summary: Register a new webhook
      description: Register a webhook URL to receive expense creation notifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
                - event
              properties:
                url:
                  type: string
                  format: uri
                  example: https://example.com/webhook
                event:
                  type: string
                  enum: [expense.created]
                  example: expense.created
                secret:
                  type: string
                  description: Optional secret for HMAC signature verification
      responses:
        '201':
          description: Webhook registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  webhook:
                    $ref: '#/components/schemas/Webhook'
  /api/webhooks/{id}:
    delete:
      tags:
        - Webhooks
      summary: Delete a webhook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Webhook deleted
        '404':
          description: Webhook not found
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john@example.com
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Webhook:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
          format: uri
        event:
          type: string
          enum: [expense.created]
        secret:
          type: string
          nullable: true
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: string